refactor: トップページの見た目を改修
