+++ src/ecosystem/history/index.module.scss
@@ -3,6 +3,7 @@
   flex-direction: row;
   width: 100%;
   height: 100%;
+  margin: auto;
 
   @media screen and (min-width: 1025px) {
     max-height: 500px;
