+++ src/ecosystem/member/index.module.scss
@@ -5,6 +5,7 @@
   flex-direction: row;
   width: 100%;
   height: 100%;
+  margin: auto;
 
   @media screen and (min-width: 1025px) {
     max-height: 500px;
@@ -58,18 +59,6 @@
 
 .char_content_wrapper {
   margin: 7% 0;
-}
-
-.small_thumbnail_wrapper {
-  height: 50%;
-  width: 28%;
-  margin: auto 1%;
-}
-
-.large_thubmnail_wrapper {
-  height: 100%;
-  width: 38%;
-  margin: 0 1%;
 }
 
 .scroll_wrapper {
