+++ src/ecosystem/top/index.module.scss
@@ -1,0 +1,55 @@
+@import "../../template.scss";
+
+.inner {
+  display: flex;
+  flex-direction: row;
+  width: 100%;
+  height: 100%;
+  margin: auto;
+
+  @media screen and (min-width: 1025px) {
+    max-width: 1200px;
+  }
+
+  @media screen and (max-width: 1024px) {
+    flex-direction: column;
+    padding-top: 80px;
+  }
+}
+
+.content_block {
+  height: 100%;
+  width: 80%;
+  margin: auto;
+
+  // NOTE: タブレット or スマホ画面では横一杯で表示する
+  @media screen and (max-width: 1024px) {
+    width: 100%;
+    margin: auto;
+  }
+}
+
+.scroll_wrapper {
+  text-align: center;
+}
+
+.title {
+  text-align: center;
+  font-size: min(4.5vw, 60px);
+  letter-spacing: 10px;
+  padding-bottom: 80px;
+
+  @media screen and (max-width: 1024px) {
+    font-size: 8vw;
+    letter-spacing: 5px;
+  }
+}
+
+.content {
+  text-align: center;
+  font-size: min(2.5vw, 35px);
+
+  @media screen and (max-width: 1024px) {
+    font-size: 4vw;
+  }
+}
